- Extracted TS part from LivePlotting.tsx into a separate file useLivePlotting.ts - Fixed LivePlotting unit tests for missing data due to these changes
